Move server impl to module server
